+++ packages/editor/editor-core/src/__tests__/unit/plugins/lists/index.ts
@@ -911,6 +911,82 @@ describe('lists', () => {
         expect(editorView.state.selection.$from.depth).toEqual(5);
       });
 
+      it("shouldn't increase the depth of list item when Tab key press when at 6 levels indentation", () => {
+        const { editorView } = editor(
+          doc(
+            ol(
+              li(
+                p('first'),
+                ol(
+                  li(
+                    p('second'),
+                    ol(
+                      li(
+                        p('third'),
+                        ol(
+                          li(
+                            p('fourth'),
+                            ol(
+                              li(
+                                p('fifth'),
+                                ol(li(p('sixth'), p('maybe seventh{<>}'))),
+                              ),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ),
+        );
+
+        expect(editorView.state.selection.$from.depth).toEqual(13);
+
+        sendKeyToPm(editorView, 'Tab');
+
+        expect(editorView.state.selection.$from.depth).toEqual(13);
+      });
+
+      it("shouldn't increase the depth of list item when Tab key press when a child list at 6 levels indentation", () => {
+        const { editorView } = editor(
+          doc(
+            ol(
+              li(
+                p('first'),
+                ol(
+                  li(
+                    p('second'),
+                    ol(
+                      li(
+                        p('third'),
+                        ol(
+                          li(
+                            p('fourth'),
+                            ol(
+                              li(p('fifth')),
+                              li(p('{<}fifth{>}'), ol(li(p('sixth')))),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ),
+        );
+
+        expect(editorView.state.selection.$from.depth).toEqual(11);
+
+        sendKeyToPm(editorView, 'Tab');
+
+        expect(editorView.state.selection.$from.depth).toEqual(11);
+      });
+
       it('should nest the list item when Tab key press', () => {
         const { editorView } = editor(
           doc(ol(li(p('text')), li(p('te{<>}xt')), li(p('text')))),