+++ packages/editor/editor-confluence-transformer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-cq-transformer
 
+## 5.1.2
+- Updated dependencies [e858305](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e858305):
+  - @atlaskit/editor-test-helpers@6.3.5
+  - @atlaskit/editor-common@26.0.0
+  - @atlaskit/editor-core@92.0.19
+
 ## 5.1.1
 - Updated dependencies [b3738ea](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b3738ea):
   - @atlaskit/editor-common@25.0.0